+++ src/PrinterTrait.php
@@ -76,6 +76,12 @@ trait PrinterTrait
     private $anyBarPort = null;
 
     /**
+     * @var bool
+     */
+    private $dontFormatClassName;
+
+
+    /**
      * {@inheritdoc}
      */
     public function __construct(
@@ -109,7 +115,7 @@ trait PrinterTrait
      */
     public function getConfigurationFile($configFileName = 'phpunit-printer.yml')
     {
-        $defaultConfigFilename = $this->getPackageRoot() . DIRECTORY_SEPARATOR . 'src/' .$configFileName;
+        $defaultConfigFilename = $this->getPackageRoot() . DIRECTORY_SEPARATOR . 'src/' . $configFileName;
 
         $configPath = getcwd();
         $filename   = '';
@@ -280,6 +286,7 @@ trait PrinterTrait
         $this->hideNamespace = $this->getConfigOption('cd-printer-hide-namespace');
         $this->anyBarEnabled = $this->getConfigOption('cd-printer-anybar');
         $this->anyBarPort    = $this->getConfigOption('cd-printer-anybar-port');
+        $this->dontFormatClassName = $this->getConfigOption('cd-printer-dont-format-classname');
 
         $this->markers = [
             'pass'         => $this->getConfigMarker('cd-pass'),
@@ -379,6 +386,11 @@ trait PrinterTrait
         if ($this->hideNamespace && strrpos($className, '\\')) {
             $className = substr($className, strrpos($className, '\\') + 1);
         }
+
+        if ($this->dontFormatClassName) {
+            return $prefix . $className . $suffix;
+        }
+
         $formattedClassName = $prefix . $className . $suffix;
 
         if (\strlen($formattedClassName) <= $this->maxClassNameLength) {